fix(platform): resolve tokenizer and scrubber patterns from one walk
createTokenizer materialized its own pattern list with a private copy of
the scrubber's factory walk — and that copy skipped the admin's custom
patterns, so a policy in tokenize mode masked an employee-id pattern at
the indexing gate but let the same value through when the chat guardrail
tokenized a prompt. The two entry points also drifted in how they
reported an unknown pattern name.

`materializePatterns` (lib/pii/engine/materialize.ts) is now the one
resolution from ScrubberOptions to the compiled list — locale union,
enabled factories in toggle order, then vetted custom patterns — and
both createScrubber and createTokenizer consume it. The tokenizer's
public surface (createTokenizer / detokenize / TokenEntry) is unchanged:
the chat guardrail's tokenize filter restores model output through it,
so the mode's promise is backed by a real round trip.

Finding: lib-utils-pii-3.
